Add optional ANGLE support for Windows destop.
Originally From https://github.com/EIRTeam/godot.git
Some precision fixes by iFire at https://github.com/V-Sekai/godot.git

Put the self-compiled angle folder in thirdparty. The release folder should be `thirdparty/angle/out/Release`.

+++ drivers/gles3/shaders/scene.glsl
@@ -114,7 +114,7 @@ uniform highp mat4 world_transform;
 layout(std140) uniform DirectionalLightData { //ubo:3
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -134,7 +134,7 @@ layout(std140) uniform DirectionalLightData { //ubo:3
 struct LightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -752,7 +752,7 @@ layout(std140) uniform SceneData {
 layout(std140) uniform DirectionalLightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -777,7 +777,7 @@ in vec4 specular_light_interp;
 struct LightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
